Post MSVC fixes clean-up
