+++ test/range.cpp
@@ -56,50 +56,6 @@ namespace tts
   };
 }
 
-namespace tts
-{
-
-  template<typename T, typename Distribution>
-  struct prng_generator
-  {
-    template<typename U, typename V>
-    prng_generator(U mn, V mx) : distribution_( static_cast<T>(mn)
-                                              , static_cast<T>(mx)
-                                              )
-    {}
-
-    void init( options const& args )
-    {
-      auto seed = args.value_or(-1, "-s", "--seed");
-
-      if(seed == -1 )
-      {
-        auto now  = std::chrono::high_resolution_clock::now();
-        seed      = static_cast<unsigned int>(now.time_since_epoch().count());
-      }
-
-      generator_.seed(seed);
-    }
-
-    template<typename Idx, typename Count>
-    T operator()(Idx, Count)
-    {
-      return distribution_(generator_);
-    }
-
-    Distribution  distribution_;
-    std::mt19937  generator_;
-  };
-
-  template<typename T>
-  using uniform_prng = prng_generator < T
-                                      , std::conditional_t< std::is_floating_point_v<T>
-                                                          , std::uniform_real_distribution<T>
-                                                          , std::uniform_int_distribution<T>
-                                                          >
-                                      >;
-}
-
 TTS_CASE( "Test stateless range check" )
 {
   TTS_ULP_RANGE_CHECK ( [](auto i, auto c) { return (100.f*i)/c; }
